+++ packages/webview_flutter/webview_flutter_wkwebview/ios/Classes/FLTWebViewFlutterPlugin.m
@@ -4,6 +4,7 @@
 
 #import "FLTWebViewFlutterPlugin.h"
 #import "FLTCookieManager.h"
+#import "FWFWebViewHostApi.h"
 #import "FlutterWebView.h"
 
 @implementation FLTWebViewFlutterPlugin